refactor : MemberTemplateEntity를 aggregate 패키지로 이동
